lift `fetchAccountDetails` to make the data more accesible
